Remove TODO and fix warning left by previous cleanup
A previous commit (#a95d776) made the variable `file_size` unused, so I removed it.
The TODO about testing with a range of WebP files has been removed because we aren't supporting
images that have no content, and that was the only type of file I still hadn't tested.

Also, we are now explicitly not supporting WebP files smaller than 30 bytes.