Fix memory leaks in image.c
The if statement in load_webp_frames which handled loading only the first frame did not destroy the
WebPAnimDecoder object or free the bytes loaded from the image.

There is still a memory issue where the font is not closed in window.c but I have a one-line fix
for that which can be put into its own PR (it's a one-line fix and should be merged easily).