Add ability to set default cli args from the config file
The cli defaults can now be set from config.h. The current defaults are
the same. This feature was requested before and shutdown because of the
existence of wrappers that could provide such default args. The main
benefit is just move towards a world where all the configuration can be
done completely in config.h.

To take advantage of this, put something like this in config.h:
```
extern opt_t _options;
__attribute__((constructor)) void constructor () {
    _options.hide_bar = true;
}
```