include config.h after all declarations
in dwm [0] as well as dmenu [1] config.h is included after all
declarations so that nested code can access those variables.
since we've changed how keybindings are handled in 12efa0e , allowing
users to write functions inside config.h, it makes sense to move the
config.h to be #included at the bottom.

[0]: https://git.suckless.org/dwm/file/dwm.c.html#l272
[1]: https://git.suckless.org/dmenu/file/dmenu.c.html#l56