Fully customizable window title
This partially undoes the change introduces in #23 as TITLE_SUFFIX has
been removed. In its place is a function pointer that accepts the
current file path and returns the title. With this the user could
customize the title however they wanted including passing the result to
a shell program or parsing printf like flags.

The '-T' option has been kept so the user can still influence the title
from the cli and the default title hasn't changed.

Closes #97