fix: reset statusbar after failed keyhandler
currently if the keyhandler invokation fails, for example due to it not
being present, the statusbar does not reset and stays on "getting
keyhandler input" message.

now the return value from run_key_handler() is used to determine if the
function was successful or not. and if the function failed, we call
handle_key_handler with false which resets the statusbar.