defer realpath() canonicalisation until access
During startup, we trigger a realpath(3) call for every argument in
order to generate absolute paths before the window opens. This might
work fine on a fast drive, but on a slow drive, forcing immediate
directory traversal and symlink resolution for the entire queue can
cause things to hang for several minutes.

Let's make the path handling more lazy. Compared to loading the image,
doing realpath(3) on load adds no cost, and improves startup time even
in the fast disk case.

inotify/etc functionality is preserved, the heavy I/O cost is simply
moved from a huge upfront cost to pay on startup to a small cost to pay
when loading the image.