fix: initialise oldst on stat failure
We save file timestamps before running the key handler and compare them
afterwards to detect changes.

As NRK pointed out, there's a bug in the existing code here: stat(2)
does not specify the state of the buffer when it fails. Zero out the
structure on failure so that the comparison will detect a change and
force a reload.

Reported-by: NRK <nrk@disroot.org>